add typescript background script
